Show title for add filter links in ListReport and BreakdownModal
